renamed generator module to ir
