+++ src/lib.rs
@@ -1,6 +1,6 @@
 pub mod driver;
 pub mod parser;
 pub mod symbol_map;
-mod generator;
+mod ir;
 
 pub use symbol_map::SymbolMap;
